Make custom commands work with multiple editors.
Previously the argument which CodeMirror passes to a command was ignored and
the handler always received the instance of the editor which registered the
command. This made it impossible to use custom commands with more than
one editor.
Now the instance of the editor which corresponds to the supplied argument is
retrieved before the handler is invoked.